Close the remaining low-severity issues and bounds-check mapper links (#9, #11-#17)
Remove unused-symbol warnings in fac_servo and fac_imu

Mark the mix/function id constants as intentionally unused, templates included

+++ Core/Src/FAC_Code/fac_battery.c
@@ -62,7 +62,7 @@ uint16_t FAC_battery_GET_cell_voltage() {
  * @bief 	Return the calculated battery type
  * @retval 	Return the calculate battery type enum value (1,2,3,4S, USB, UNKNOWN)
  */
-uint16_t FAC_battery_GET_type(uint16_t vbat) {
+uint8_t FAC_battery_GET_type(uint16_t vbat) {
 	FAC_battery_calculate_type(vbat);
 	return battery.type;
 }
